requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Werkzeug==3.0.3
PyMuPDF==1.24.9
pandas==2.2.2
python-docx==1.1.2
openpyxl==3.1.5
xlsxwriter==3.2.0
